Add science category (#5)
* menu: change "Education" to "Education & Science"

Current menus do not cover the `Science` category. It is generally not
an issue as most applications that have category `Science` (ex:
Libreoffice Math, GeoGebra) also have category `Education`. However,
some applications (ex: Picoscope [1]) have cateogory `Science` only and
not `Education`, and thus they are classified as "Other".

A considered fix is adding an "Education" submenu using
`lxqt-education.directory`. However, that directory uses the same icon
(`applications-science`) as `lxqt-science-math.directory`, and it cannot
switch to `applications-education` icon, as the latter is not a standard
icon name [2] and is missing from many icon themes [3]. As a result,
users will see two submenus with the same icon and many common
applications.

As we have full control of menu-related files after relevant LXQt
components are migrated [4], we can change menu files to include the
`Science` category.

`lxqt-education.directory` is preserved for backward compatibility. It
may be used in some custom menu files.

[1] https://www.picotech.com/
[2] https://gitlab.freedesktop.org/xdg/default-icon-theme/-/blob/master/spec/icon-naming-spec.xml
[3] https://commons.wikimedia.org/wiki/Comparison_of_icon_sets#Categories
[4] https://github.com/lxqt/lxqt/issues/1679

* Setup translations